Extract reusable code
